* feature/swift2_array_datasouce:
  File updates
  Fixed bug with app crashing when click in text field to edit drink name
  Updates
  no message
  updates
  Final revision before submitting
  Add ability to add new brand to the list
  • using arrays for the datasource

+++ Coffee Timer/TimerListTableViewController.swift
@@ -20,6 +20,28 @@ class TimerListTableViewController: UITableViewController {
     
     var userReorderingCells = false
     let cellIdentifier = "Cell"
+    var creatingNewTimer = false
+    
+    var _coffees: [TimerModel] {
+        let request = NSFetchRequest(entityName: "TimerModel")
+        request.predicate = NSPredicate(format: "type == %d", TableSection.Coffee.rawValue)
+        request.sortDescriptors = [NSSortDescriptor(key: "name", ascending: true, selector: "caseInsensitiveCompare:")]
+        let coffees: [TimerModel] = try! appDelegate().coreDataStack.managedObjectContext.executeFetchRequest(request) as! [TimerModel]
+        
+        return coffees
+    }
+    
+    var _teas: [TimerModel] {
+        let request = NSFetchRequest(entityName: "TimerModel")
+        request.predicate = NSPredicate(format: "type == %d", TableSection.Tea.rawValue)
+        request.sortDescriptors = [NSSortDescriptor(key: "name", ascending: true, selector: "caseInsensitiveCompare:")]
+        let teas: [TimerModel] = try! appDelegate().coreDataStack.managedObjectContext.executeFetchRequest(request) as! [TimerModel]
+        
+        return teas
+    }
+    
+    var coffees: [TimerModel]?
+    var teas: [TimerModel]?
     
     lazy var fetchedResultsController: NSFetchedResultsController = {
         let fetchRequest = NSFetchRequest(entityName: "TimerModel")
@@ -38,9 +60,22 @@ class TimerListTableViewController: UITableViewController {
         case NumberOfSections
     }
     
-    func timerModelForIndexPath(indexPath: NSIndexPath) -> TimerModel
-    {
-        return fetchedResultsController.objectAtIndexPath(indexPath) as! TimerModel
+    func timerModelForIndexPath(indexPath: NSIndexPath) -> TimerModel? {
+        var timerModel: TimerModel?
+        
+        switch indexPath.section {
+            case TableSection.Coffee.rawValue:
+                if self.coffees!.count > 0 {
+                    timerModel = coffees![indexPath.row]
+                }
+            case TableSection.Tea.rawValue:
+                if self.teas!.count > 0 {
+                    timerModel = teas![indexPath.row]
+                }
+            default: return timerModel
+        }
+        
+        return timerModel
     }
     
     override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?)
@@ -51,26 +86,32 @@ class TimerListTableViewController: UITableViewController {
             
             if segue.identifier == "pushDetail" {
                 let detailViewController = segue.destinationViewController as! TimerDetailViewController
+                
                 detailViewController.timerModel = timerModel
-            } else if segue.identifier == "editDetail" {
+            }
+            else if segue.identifier == "editDetail" {
+                print("editDetail prepareForSegue")
                 let navigationController = segue.destinationViewController as! UINavigationController
                 let editViewController = navigationController.topViewController as! TimerEditViewController
                 
                 editViewController.timerModel = timerModel
                 editViewController.delegate = self
             }
-        } else if segue.identifier == "newTimer" {
-                let navigationController = segue.destinationViewController as! UINavigationController
-                let editViewController = navigationController.topViewController as! TimerEditViewController
-                
-                editViewController.creatingNewTimer = true
-                editViewController.timerModel = NSEntityDescription.insertNewObjectForEntityForName("TimerModel", inManagedObjectContext: appDelegate().coreDataStack.managedObjectContext) as! TimerModel
-                editViewController.delegate = self
+        }
+        else if segue.identifier == "newTimer" {
+            let navigationController = segue.destinationViewController as! UINavigationController
+            let editViewController = navigationController.topViewController as! TimerEditViewController
+            
+            editViewController.creatingNewTimer = true
+            creatingNewTimer = true
+            editViewController.timerModel = NSEntityDescription.insertNewObjectForEntityForName("TimerModel", inManagedObjectContext: appDelegate().coreDataStack.managedObjectContext) as! TimerModel
+            editViewController.delegate = self
         }
     }
     
     override func shouldPerformSegueWithIdentifier(identifier: String?, sender: AnyObject?) -> Bool
     {
+        print("editDetail shouldPerformSegueWithIdentifier")
         if identifier == "pushDetail" {
             if tableView.editing {
                 return false
@@ -90,23 +131,35 @@ class TimerListTableViewController: UITableViewController {
             print("Error fetching: \(error)")
         }
         title = "Drinks"
-//        let tabBarIndex = self.navigationController?.tabBarController?.selectedIndex
         navigationItem.leftBarButtonItem = editButtonItem()
         self.tableView.contentInset = UIEdgeInsetsMake(0, 0, 44.0, 0)
+        
+        coffees = _coffees
+        teas = _teas
     }
     
     override func viewWillAppear(animated: Bool)
     {
         super.viewWillAppear(animated)
-        if presentedViewController != nil {
-            tableView.reloadData()
-        }
+        tableView.reloadData()
+
+    }
+    
+    override func viewDidAppear(animated: Bool) {
+        super.viewDidAppear(animated)
     }
     
     override func setEditing(editing: Bool, animated: Bool)
     {
         super.setEditing(editing, animated: animated)
         navigationItem.rightBarButtonItem?.enabled = !editing
+        self.shouldDisableBarButtonItems(!editing)
+    }
+    
+    func shouldDisableBarButtonItems(enabled: Bool) {
+        for item in self.tabBarController!.tabBar.items! {
+            item.enabled = enabled
+        }
     }
 
     // MARK: - Table view data source
@@ -117,67 +170,82 @@ class TimerListTableViewController: UITableViewController {
         return TableSection.NumberOfSections.rawValue
     }
     
-    override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath)
+    override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int
     {
+        switch section {
+            case TableSection.Coffee.rawValue: return coffees!.count
+            case TableSection.Tea.rawValue: return teas!.count
+            default: return 0
+        }
+    }
+
+    override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+        print("in didSelectRowAtIndexPath")
         if tableView.editing {
             let cell = tableView.cellForRowAtIndexPath(indexPath)
             performSegueWithIdentifier("editDetail", sender: cell)
         }
     }
-
-    override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int
-    {
-        if section >= fetchedResultsController.sections?.count {
-            return 0
-        }
-        guard let sectionInfo: NSFetchedResultsSectionInfo = (fetchedResultsController.sections?[section])! else {
-            return 0
-        }
-        return sectionInfo.numberOfObjects
-    }
-
-    override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell
-    {
+    
+    override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCellWithIdentifier(cellIdentifier, forIndexPath: indexPath)
-        let timerModel = timerModelForIndexPath(indexPath)
-        cell.textLabel?.text = timerModel.name
-        if let brand = timerModel.brand as BrandModel? {
-            cell.detailTextLabel?.text = brand.name
+        if let timerModel: TimerModel = timerModelForIndexPath(indexPath) {
+            cell.textLabel?.text = timerModel.name
+            if let brand = timerModel.brand as BrandModel? {
+                cell.detailTextLabel?.text = brand.name
+            }
         }
         
         return cell
     }
     
-    override func tableView(tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int)
-    {
+    override func tableView(tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int) {
         let headerView: UITableViewHeaderFooterView = (view as? UITableViewHeaderFooterView)!
         headerView.contentView.backgroundColor = UIColor(red: 0.8, green: 0.95, blue: 1, alpha: 0.5)
     }
     
-    override func tableView(tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat
-    {
-        if section < fetchedResultsController.sections?.count {
-            let sectionInfo: NSFetchedResultsSectionInfo = (fetchedResultsController.sections?[section])!
-            return sectionInfo.numberOfObjects > 0 ? 44 : 0
+    override func tableView(tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        switch section {
+            case TableSection.Coffee.rawValue: return coffees!.count > 0 ? 44 : 0
+            case TableSection.Tea.rawValue: return teas!.count > 0 ? 44 : 0
+            default: return 0
         }
-        return 0
+//        if section < fetchedResultsController.sections?.count {
+//            let sectionInfo: NSFetchedResultsSectionInfo = (fetchedResultsController.sections?[section])!
+//            return sectionInfo.numberOfObjects > 0 ? 44 : 0
+//        }
+//        return 0
     }
 
-    override func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle, forRowAtIndexPath indexPath: NSIndexPath)
-    {
+    override func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle, forRowAtIndexPath indexPath: NSIndexPath) {
         if editingStyle == .Delete {
-            let timer = timerModelForIndexPath(indexPath)
-            timer.managedObjectContext?.deleteObject(timer)
+            if let timer = timerModelForIndexPath(indexPath) {
+                removeItemFromDataSource(indexPath)
+                timer.managedObjectContext?.deleteObject(timer)
+                appDelegate().saveCoreData()
+            }
+        }
+    }
+    
+    func removeItemFromDataSource(atIndexPath: NSIndexPath) {
+        switch atIndexPath.section {
+            case TableSection.Coffee.rawValue:
+                if coffees!.count > 0 {
+                    coffees!.removeAtIndex(atIndexPath.row)
+                }
+            case TableSection.Tea.rawValue:
+                if teas!.count > 0 {
+                    teas!.removeAtIndex(atIndexPath.row)
+                }
+            default: return
         }
     }
         
-    override func tableView(tableView: UITableView, shouldShowMenuForRowAtIndexPath indexPath: NSIndexPath) -> Bool
-    {
+    override func tableView(tableView: UITableView, shouldShowMenuForRowAtIndexPath indexPath: NSIndexPath) -> Bool {
         return true
     }
     
-    override func tableView(tableView: UITableView, canPerformAction action: Selector, forRowAtIndexPath indexPath: NSIndexPath, withSender sender: AnyObject?) -> Bool
-    {
+    override func tableView(tableView: UITableView, canPerformAction action: Selector, forRowAtIndexPath indexPath: NSIndexPath, withSender sender: AnyObject?) -> Bool {
         if action == "copy:" {
             return true
         }
@@ -185,16 +253,15 @@ class TimerListTableViewController: UITableViewController {
         return false
     }
 
-    override func tableView(tableView: UITableView, performAction action: Selector, forRowAtIndexPath indexPath: NSIndexPath, withSender sender: AnyObject?)
-    {
-        let timerModel = timerModelForIndexPath(indexPath)
-        let pasteboard = UIPasteboard.generalPasteboard()
-        pasteboard.string = timerModel.name
+    override func tableView(tableView: UITableView, performAction action: Selector, forRowAtIndexPath indexPath: NSIndexPath, withSender sender: AnyObject?) {
+        if let timerModel = timerModelForIndexPath(indexPath) {
+            let pasteboard = UIPasteboard.generalPasteboard()
+            pasteboard.string = timerModel.name
+        }
     }
     
-    override func tableView(tableView: UITableView, targetIndexPathForMoveFromRowAtIndexPath sourceIndexPath: NSIndexPath, toProposedIndexPath proposedDestinationIndexPath: NSIndexPath) -> NSIndexPath
-    {
-       // If the source and destination index paths are the same section,
+    override func tableView(tableView: UITableView, targetIndexPathForMoveFromRowAtIndexPath sourceIndexPath: NSIndexPath, toProposedIndexPath proposedDestinationIndexPath: NSIndexPath) -> NSIndexPath {
+        // If the source and destination index paths are the same section,
         // then return the proposed index path
         if sourceIndexPath.section == proposedDestinationIndexPath.section {
             return proposedDestinationIndexPath
@@ -218,32 +285,32 @@ class TimerListTableViewController: UITableViewController {
         }
     }
     
-    override func tableView(tableView: UITableView, moveRowAtIndexPath sourceIndexPath: NSIndexPath, toIndexPath destinationIndexPath: NSIndexPath)
-    {
-        userReorderingCells = true
-        
-        // Grab the section and the TimerModels in the section
-        let sectionInfo: NSFetchedResultsSectionInfo = (fetchedResultsController.sections?[sourceIndexPath.section])!
-        
-        var objectsInSection = sectionInfo.objects ?? []
-        
-        // Rearrange the order to match the user's actions
-        // Note: this doesn't move anything in Core Data, just our objectsInSection array
-        objectsInSection.moveFrom(sourceIndexPath.row, toDestination: destinationIndexPath.row)
-        
-        // The models are now in the correct order.
-        // Update their displayOrder to match the new order.
-        for i in 0..<objectsInSection.count {
-            let model = objectsInSection[i] as? TimerModel
-            model?.displayOrder = Int32(i)
-        }
-        
-        userReorderingCells = false
-        do {
-            try appDelegate().coreDataStack.managedObjectContext.save()
-        } catch _ {
-        }
-    }
+//    override func tableView(tableView: UITableView, moveRowAtIndexPath sourceIndexPath: NSIndexPath, toIndexPath destinationIndexPath: NSIndexPath)
+//    {
+//        userReorderingCells = true
+//        
+//        // Grab the section and the TimerModels in the section
+//        let sectionInfo: NSFetchedResultsSectionInfo = (fetchedResultsController.sections?[sourceIndexPath.section])!
+//        
+//        var objectsInSection = sectionInfo.objects ?? []
+//        
+//        // Rearrange the order to match the user's actions
+//        // Note: this doesn't move anything in Core Data, just our objectsInSection array
+//        objectsInSection.moveFrom(sourceIndexPath.row, toDestination: destinationIndexPath.row)
+//        
+//        // The models are now in the correct order.
+//        // Update their displayOrder to match the new order.
+//        for i in 0..<objectsInSection.count {
+//            let model = objectsInSection[i] as? TimerModel
+//            model?.displayOrder = Int32(i)
+//        }
+//        
+//        userReorderingCells = false
+//        do {
+//            try appDelegate().coreDataStack.managedObjectContext.save()
+//        } catch _ {
+//        }
+//    }
     
     override func tableView(tableView: UITableView, titleForHeaderInSection section: Int) -> String?
     {
@@ -256,45 +323,75 @@ class TimerListTableViewController: UITableViewController {
 }
 
 extension TimerListTableViewController: NSFetchedResultsControllerDelegate {
-    func controllerWillChangeContent(controller: NSFetchedResultsController) {
-        tableView.beginUpdates()
-    }
+//    func controllerWillChangeContent(controller: NSFetchedResultsController) {
+//        tableView.beginUpdates()
+//    }
+//    
+//    func controllerDidChangeContent(controller: NSFetchedResultsController) {
+//        tableView.endUpdates()
+//    }
     
-    func controllerDidChangeContent(controller: NSFetchedResultsController) {
-        tableView.endUpdates()
-    }
-    
-    func controller(controller: NSFetchedResultsController, didChangeObject anObject: AnyObject, atIndexPath indexPath: NSIndexPath?, forChangeType type: NSFetchedResultsChangeType, newIndexPath: NSIndexPath?)
-    {
+//    func controller(controller: NSFetchedResultsController, didChangeObject anObject: AnyObject, atIndexPath indexPath: NSIndexPath?, forChangeType type: NSFetchedResultsChangeType, newIndexPath: NSIndexPath?)
+//    {
+//        if userReorderingCells {
+//            return
+//        }
+//        
+//        switch type {
+//        case .Insert:
+//            return
+//            tableView.insertRowsAtIndexPaths([newIndexPath!], withRowAnimation: .Automatic)
+//        case .Delete:
+//            tableView.deleteRowsAtIndexPaths([indexPath!], withRowAnimation: .Automatic)
+//        case .Move:
+//            return
+//            tableView.moveRowAtIndexPath(indexPath!, toIndexPath: newIndexPath!)
+//        case .Update:
+//            tableView.reloadRowsAtIndexPaths([indexPath!], withRowAnimation: .Automatic)
+//        }
+//
+//        tableView.reloadData()
+//    }
+    func controller(controller: NSFetchedResultsController, didChangeObject anObject: AnyObject, atIndexPath indexPath: NSIndexPath?, forChangeType type: NSFetchedResultsChangeType, newIndexPath: NSIndexPath?) {
+        
         if userReorderingCells {
             return
         }
-        
-        switch type {
-        case .Insert:
-            tableView.insertRowsAtIndexPaths([newIndexPath!], withRowAnimation: .Automatic)
-        case .Delete:
-            tableView.deleteRowsAtIndexPaths([indexPath!], withRowAnimation: .Automatic)
-        case .Move:
-            tableView.moveRowAtIndexPath(indexPath!, toIndexPath: newIndexPath!)
-        case .Update:
-            tableView.reloadRowsAtIndexPaths([indexPath!], withRowAnimation: .Automatic)
-        }
 
-        tableView.reloadData()
+        switch type {
+        case .Delete:
+            if !creatingNewTimer {
+//                switch indexPath!.section {
+//                case TableSection.Coffee.rawValue:
+//                    coffees?.removeAtIndex(indexPath!.row)
+//                case TableSection.Tea.rawValue:
+//                    teas?.removeAtIndex(indexPath!.row)
+//                default:
+//                    return
+//                }
+                tableView.deleteRowsAtIndexPaths([indexPath!], withRowAnimation: .Automatic)
+                tableView.reloadData()
+            }
+        default:
+            return
+        }
     }
 }
 
 extension TimerListTableViewController: TimerEditViewControllerDelegate {
     func timerEditViewControllerDidCancel(viewController: TimerEditViewController)
     {
-        if viewController.creatingNewTimer {
-            appDelegate().coreDataStack.managedObjectContext.deleteObject(viewController.timerModel)
-        }
+        coffees = _coffees
+        teas = _teas
+        tableView.reloadData()
+        creatingNewTimer = false
     }
     
-    func timerEditViewControllerDidSave(viewController: TimerEditViewController)
-    {
+    func timerEditViewControllerDidSave(viewController: TimerEditViewController) {
         appDelegate().saveCoreData()
+        coffees = _coffees
+        teas = _teas
+        tableView.reloadData()
+        creatingNewTimer = false
     }
 }
